# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionItemType": {
        "description": "Type of action derived from a document.",
        "enum": [
          "deadline",
          "court_date",
          "response_required",
          "filing_required",
          "payment_due"
        ],
        "title": "ActionItemType",
        "type": "string"
      },
...
      "CaseCard": {
        "description": "One case as shown on Home and in the Cases directory.",
        "properties": {
          "client_name": {
            "title": "Client Name",
            "type": "string"
          },
          "days_since_activity": {
            "title": "Days Since Activity",
            "type": "integer"
          },
          "doc_count": {
            "title": "Doc Count",
            "type": "integer"
          },
          "exposure_eur": {
            "title": "Exposure Eur",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_dormant": {
            "title": "Is Dormant",
            "type": "boolean"
          },
          "is_draft": {
            "title": "Is Draft",
            "type": "boolean"
          },
          "last_activity_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity At"
          },
          "matter_type": {
            "title": "Matter Type",
            "type": "string"
          },
          "max_significance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignificanceTier"
              },
              {
                "type": "null"
              }
            ]
          },
          "new_docs": {
            "title": "New Docs",
            "type": "integer"
          },
          "next_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NextAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "open_action_count": {
            "title": "Open Action Count",
            "type": "integer"
          },
          "opposing_party": {
            "title": "Opposing Party",
            "type": "string"
          },
          "pending_close": {
            "title": "Pending Close",
            "type": "boolean"
          },
          "proceeding_name": {
            "title": "Proceeding Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "status_label": {
            "title": "Status Label",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "status_label",
          "is_draft",
          "pending_close",
          "client_name",
          "opposing_party",
          "proceeding_name",
          "matter_type",
          "next_action",
          "exposure_eur",
          "doc_count",
          "open_action_count",
          "new_docs",
          "days_since_activity",
          "is_dormant",
          "max_significance",
          "last_activity_at"
        ],
        "title": "CaseCard",
        "type": "object"
      },
      "CaseCreate": {
        "properties": {
          "case_id": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Case Id",
            "type": "string"
          },
          "court_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Court Name",
            "type": "string"
          },
          "jurisdiction": {
            "$ref": "#/components/schemas/Jurisdiction",
            "default": "de"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "title",
          "court_name"
        ],
        "title": "CaseCreate",
        "type": "object"
      },
      "CaseCreated": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "CaseCreated",
        "type": "object"
      },
      "CaseRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_draft": {
            "title": "Is Draft",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "is_draft"
        ],
        "title": "CaseRef",
        "type": "object"
      },
      "CaseStatus": {
        "enum": [
          "intake",
          "discovery",
          "pre_trial",
          "trial",
          "post_trial",
          "closed"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
      "CasesDirectory": {
        "properties": {
          "cases": {
            "items": {
              "$ref": "#/components/schemas/CaseCard"
            },
            "title": "Cases",
            "type": "array"
          },
          "counts_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/CaseStatus"
            },
            "title": "Counts By Status",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "cases",
          "counts_by_status",
          "total"
        ],
        "title": "CasesDirectory",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Body of every non-2xx ``/api/v1`` response.\n\n``code`` is a stable machine-readable identifier the client branches on;\n``detail`` is the human-readable message it may show as-is.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Jurisdiction": {
        "description": "Case jurisdiction for cost system.",
        "enum": [
          "de",
          "uk",
          "us",
          "other"
        ],
        "title": "Jurisdiction",
        "type": "string"
      },
...
      "NextAction": {
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionItemType"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "due_date",
          "action_type"
        ],
        "title": "NextAction",
        "type": "object"
      },
...
      "SignificanceTier": {
        "description": "Document significance to case \u2014 drives graph visibility.",
        "enum": [
          "critical",
          "significant",
          "informational",
          "administrative"
        ],
        "title": "SignificanceTier",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Sanctuary API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "operationId": "v1_cases_directory_api_v1_cases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasesDirectory"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Cases Directory",
        "tags": [
          "cases"
        ]
      },
      "post": {
        "description": "Create a case and its initial active proceeding.",
        "operationId": "v1_create_case_api_v1_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Create Case",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/api/v1/cases/{case_id}/confirm-draft": {
      "post": {
        "description": "Ratify an AI-created draft case.",
        "operationId": "v1_confirm_draft_api_v1_cases__case_id__confirm_draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRef"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Confirm Draft",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/cases/{case_id}/reject-draft": {
      "post": {
        "description": "Delete an AI-created draft case; its documents return to triage.",
        "operationId": "v1_reject_draft_api_v1_cases__case_id__reject_draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Reject Draft",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}